inline sq check in prefetch instruction

+++ src/jit/frontend/armv3/armv3_frontend.h
@@ -12,7 +12,7 @@ struct armv3_frontend {
   /* runtime interface */
   void *data;
   void (*translate)(void *, uint32_t, struct ir *, int);
-  void (*switch_mode)(void *, uint64_t);
+  void (*switch_mode)(void *, uint32_t);
   void (*restore_mode)(void *);
   void (*software_interrupt)(void *);
 };
